flask-mysql-nginx/project/requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.1.1
Werkzeug==2.2.2
python-dotenv==1.0.1